@@ -1,44 +1,3 @@
-const {app, BrowserWindow} = require('electron');
-const url = require('url');
+const bootstrap = require('./bootstrap');
 
-let window;
-
-function createWindow () {
-  window = new BrowserWindow({
-    width: 1400,
-    height: 800,
-    webPreferences: {
-      nodeIntegration: true,
-      webSecurity: false
-    }
-  });
-
-  window.webContents.session.webRequest.onBeforeSendHeaders((details, callback) => {
-    delete details.requestHeaders['User-Agent'];
-    callback({cancel: false, requestHeaders: details.requestHeaders});
-  });
-
-  window.loadURL(
-    url.format({
-      pathname: 'localhost:4300',
-      protocol: 'http:',
-      slashes: true
-    })
-  );
-
-  window.webContents.openDevTools();
-
-  window.on('closed', () => {
-    window = null;
-  });
-}
-
-app.on('ready', createWindow);
-
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') app.quit();
-});
-
-app.on('activate', () => {
-  if (window === null) createWindow();
-});
+bootstrap(true);
